+++ flow/layers/default_layer_builder.h
@@ -65,7 +65,7 @@ class DefaultLayerBuilder final : public LayerBuilder {
   // |flow::LayerBuilder|
   void PushTexture(const SkPoint& offset,
                    const SkSize& size,
-                   size_t texture_id) override;
+                   int64_t texture_id) override;
 
 #if defined(OS_FUCHSIA)
   // |flow::LayerBuilder|
